"Merged" Pawn from Steves Repository
I couldn't get the GUI to do a proper fetch and merge so I just copied
the file. I fixed pawn and added the ability to replace it after it
hits the end of the board. I made changes to moveAndEndActivePlayerTurn
of ChessGame and the MoveTo method of ChessPiece. It now throws a
PieceNeedsToBeReplacedException.